+++ CHANGELOG.md
@@ -10,6 +10,7 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Added
 
 - Add `MixUp` augmentation for LT-DETR object detection training.
+- Add support for COCO object detection dataset format.
 
 ### Changed